Timeline materials
Folder with the HTML, JSON, JavaScript, as well as the HTML for the biography. CSS linked to both the biography and timeline. Events can easily be added to the timeline via the JSON file, and the timeline styling can be edited via the CSS that's ultimately linked to that page (still figuring that last part out as far as the style edits actually showing in the timeline, but can make changes to the captions and text once we decide on a standard stylesheet).